+++ tests/dotnet-config-kit.Benchmarks/obj/Release/net8.0/dotnet-config-kit.Benchmarks.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("dotnet-config-kit.Benchmarks")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Release")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+2af93071c1a8b65ed42fb29aa8bf8ea52c2b2e05")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+e0faefcf41a675d280b422b4711e0e2a7b401d5f")]
 [assembly: System.Reflection.AssemblyProductAttribute("dotnet-config-kit.Benchmarks")]
 [assembly: System.Reflection.AssemblyTitleAttribute("dotnet-config-kit.Benchmarks")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]